Reading Int safely and defaulting to -1 when failed.

+++ driver/src/main/scala/api/commands/instanceadministration.scala
@@ -53,8 +53,8 @@ case class CollStats(scale: Option[Int] = None) extends CollectionCommand with C
  * @param paddingFactor Padding can speed up updates if documents grow (only for mmapv1 storage engine).
  * @param systemFlags System flags.
  * @param userFlags User flags.
- * @param totalIndexSize The total size in bytes (or in bytes / scale) of all indexes.
- * @param sizePerIndex Size of specific indexes in bytes (or in bytes / scale).
+ * @param totalIndexSize The total size in bytes (or in bytes / scale) of all indexes. -1 if value exceeds Int.MaxValue
+ * @param sizePerIndex Size of specific indexes in bytes (or in bytes / scale). -1 if value exceeds Int.MaxValue
  * @param capped States if this collection is capped.
  * @param max The maximum number of documents of this collection, if capped.
  * @param maxSize The maximum size in bytes (or in bytes / scale, if any) of this collection, if capped.
@@ -71,13 +71,13 @@ case class CollStatsResult(
     paddingFactor: Option[Double],
     systemFlags: Option[Int],
     userFlags: Option[Int],
-    totalIndexSize: Double,
-    sizePerIndex: List[(String, Double)],
+    totalIndexSize: Int,
+    sizePerIndex: List[(String, Int)],
     capped: Boolean,
     max: Option[Long],
     maxSize: Option[Double] = None
 ) {
-  @inline def indexSizes: Array[(String, Double)] = sizePerIndex.toArray
+  @inline def indexSizes: Array[(String, Int)] = sizePerIndex.toArray
 
   @deprecated(message = "Use [[copy]] with [[maxSize]]", since = "0.11.10")
   def copy(
@@ -92,8 +92,8 @@ case class CollStatsResult(
     paddingFactor: Option[Double] = this.paddingFactor,
     systemFlags: Option[Int] = this.systemFlags,
     userFlags: Option[Int] = this.userFlags,
-    totalIndexSize: Double = this.totalIndexSize,
-    indexSizes: Array[(String, Double)] = this.sizePerIndex.toArray,
+    totalIndexSize: Int = this.totalIndexSize,
+    indexSizes: Array[(String, Int)] = this.sizePerIndex.toArray,
     capped: Boolean = this.capped,
     max: Option[Long] = this.max
   ): CollStatsResult = CollStatsResult(
